+++ packages/ente-api/test/mocks/entities.ts
@@ -1,4 +1,11 @@
-import { UserDto, EntryDto, Roles, CreateEntryDto, SlotDto } from "ente-types";
+import {
+  UserDto,
+  EntryDto,
+  Roles,
+  CreateEntryDto,
+  SlotDto,
+  dateToIsoString
+} from "ente-types";
 import { RequestContextUser } from "../../src/helpers/request-context";
 import { Some, Maybe } from "monet";
 
@@ -6,7 +13,7 @@ const withGetDto = <T extends {}>(
   object: T
 ): T & { getDto: () => Promise<Maybe<T>> } => {
   const getDto = async () => Some(object);
-  return Object.assign(object, { getDto });
+  return { ...object, getDto };
 };
 
 const admin: UserDto & RequestContextUser = withGetDto({
@@ -15,7 +22,7 @@ const admin: UserDto & RequestContextUser = withGetDto({
   displayname: "Admin",
   email: "admin@gmail.com",
   id: "1cf900d1-2c1e-4178-907a-23e327f11d05",
-  isAdult: false,
+  birthday: undefined,
   role: Roles.ADMIN,
   username: "admin"
 });
@@ -26,7 +33,7 @@ const tomTallis: UserDto & RequestContextUser = withGetDto({
   displayname: "Thomas Tallis",
   email: "tom@tallis.de",
   id: "1cf900d1-2c1e-4178-907a-23e327f11d05",
-  isAdult: false,
+  birthday: "2000-02-01",
   role: Roles.STUDENT,
   username: "tomtallis"
 });
@@ -36,14 +43,14 @@ const benBongo: UserDto & RequestContextUser = withGetDto({
   childrenIds: [],
   displayname: "Mr. Bongo",
   email: "bongo@my-school.com",
-  id: "1cf900d1-2c1e-4178-907a-23e327f11d05",
-  isAdult: false,
+  id: "1cf900d1-2c1e-4178-907a-23e327f11d06",
+  birthday: undefined,
   role: Roles.TEACHER,
   username: "benbongo"
 });
 
 const slot: SlotDto = {
-  id: "1cf900d1-2c1e-4178-907a-23e327f11d05",
+  id: "1cf900d1-2c1e-4178-904a-23e327f11d05",
   date: new Date(),
   from: 3,
   to: 4,
@@ -54,7 +61,7 @@ const slot: SlotDto = {
 };
 
 const entry: EntryDto = {
-  id: "1cf900d1-2c1e-4178-907a-23e327f11d05",
+  id: "1cf900d1-2c1e-4178-903a-23e327f11d05",
   date: new Date(),
   createdAt: new Date(),
   updatedAt: new Date(),
@@ -66,14 +73,13 @@ const entry: EntryDto = {
 };
 
 const createEntry: CreateEntryDto = {
-  date: new Date(),
+  date: "2018-12-03",
   forSchool: false,
   slots: [
     {
       from: 1,
       teacherId: benBongo.id,
-      to: 1,
-      date: new Date()
+      to: 1
     }
   ]
 };