+++ packages/ente-types/src/date-to-iso-string.ts
@@ -1,0 +1,4 @@
+import { format } from "date-fns";
+
+export const dateToIsoString = (d: Date | number | string) =>
+  format(d, "yyyy-MM-dd");