+++ packages/ente-types/src/roles.ts
@@ -15,3 +15,10 @@ export enum Roles {
 }
 
 export const rolesArr = Object.keys(Roles).map(key => (Roles as any)[key]);
+
+export const roleHasGraduationYear = (role: Roles) =>
+  [Roles.STUDENT, Roles.MANAGER].includes(role);
+export const roleHasChildren = (role: Roles) => role === Roles.PARENT;
+export const roleTeaches = (role: Roles) =>
+  [Roles.MANAGER, Roles.TEACHER].includes(role);
+export const roleHasBirthday = (role: Roles) => role === Roles.STUDENT;